Update hw1_2.c
Given an array containing n keys, design an algorithm to determine whether there exists such a key which is equal to the summation of other two keys in the array. Explain the worst-case time complexity (no sequential search is allowed). Print out all the keys that are equal to the sum of the other two keys in the array, and print out the corresponding two keys.