Add files via upload
Implement the algorithm for finding the closest pair of points in two dimension plane using divide and conquer strategy. Use the brute-force approach to solve the above problem. Compare the two implemented algorithms in terms of time complexity. Print out the time cost during the execution of these two algorithms.